pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zephyrveil"
version = "1.0.0"
description = "Linux threat detection and security intelligence tool"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["security", "linux", "threat-detection", "HIDS", "log-analysis"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Topic :: Security",
]
dependencies = [
    "rich",
    "reportlab",
    "requests",
    "shodan",
    "packaging",
]

[project.optional-dependencies]
dev = ["python-dotenv"]

[project.scripts]
zephyrveil = "zephyrveil.main:main"

[project.urls]
Homepage = "https://github.com/avdrevygh/zephyrveil"
Repository = "https://github.com/avdrevygh/zephyrveil"
Issues = "https://github.com/avdrevygh/zephyrveil/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/zephyrveil"]
